backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
typing_extensions==4.8.0
psycopg2-binary==2.9.9
psycopg2==2.9.9
djangorestframework==3.14.0
django-cors-headers == 4.3.0
google-cloud-bigquery == 3.13.0
google-auth == 2.23.4